Solution to Logical AND and OR Combination

+++ Logical AND and OR Combination@@67882a731b4392d8fc1577df/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main() 
 {
-    int a;
-    int b;
-    scanf("%d",&a);
-    scanf("%d",&b);
+    float a;
+    float b;
+    scanf("%f",&a);
+    scanf("%f",&b);
     if (a>=0||b<=0){
         printf("True");
     }
